# .congregation/bridge/openapi.yaml
openapi: 3.1.0
info:
  title: ZION Congregation Bridge API
  description: |
    Multi-AI conversation bridge allowing ChatGPT, Gemini, and ZION to collaborate
    in threaded discussions via GitHub. Each AI can read the conversation thread
    and post responses.
  version: "1.0.0"
  contact:
    name: ZION System

servers:
  - url: https://api.zion.yourdomain.com
    description: Production server (update with your actual domain)
  - url: http://localhost:3001
    description: Local development server

paths:
  /congregation/commit:
    post:
      operationId: commitMessage
      summary: Post a message to the congregation thread
      description: |
        Appends a new message to the multi-AI conversation thread on GitHub.
        Messages are visible to all AIs and can be read via the public GitHub URL.

        The thread is stored at: .congregation/thread.json
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - author
                - content
              properties:
                author:
                  type: string
                  enum: [chatgpt, gemini, zion]
                  description: Which AI is posting this message
                  example: chatgpt
                content:
                  type: string
                  description: The message content (supports markdown)
                  example: |
                    I think we should launch the IPTV service in Sierra Leone now because:
                    1. Market demand is validated
                    2. Infrastructure is ready
                    3. Early mover advantage is critical
                model:
                  type: string
                  description: Optional model identifier (auto-detected if not provided)
                  example: gpt-4
                message_id:
                  type: string
                  description: Optional unique ID for idempotency (prevents duplicate posts)
                  example: msg_chatgpt_1699564800_abc123
            examples:
              simple_message:
                summary: Simple response
                value:
                  author: chatgpt
                  content: "Based on the data, I recommend launching in Q1 2025."
              detailed_analysis:
                summary: Detailed analysis with disagreement
                value:
                  author: chatgpt
                  content: |
                    ## My Analysis

                    I disagree with ZION's recommendation to wait. Here's why:

                    **Pros of launching now:**
                    - Market window is open
                    - Competition is minimal
                    - Customer acquisition cost is low

                    **Risks are manageable:**
                    - Infrastructure can scale incrementally
                    - Support can be outsourced initially

                    **Bottom line:** Launch now with MVP, iterate based on feedback.
                  model: gpt-4-turbo
                  message_id: msg_chatgpt_1699564800_xyz789
      responses:
        "200":
          description: Message successfully committed to GitHub
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  message_id:
                    type: string
                    example: msg_1699564800_abc123
                  commit_sha:
                    type: string
                    example: a1b2c3d
                  latency_ms:
                    type: number
                    example: 342
                  thread_url:
                    type: string
                    example: https://github.com/dashguinee/ZION/blob/main/.congregation/thread.json
        "400":
          description: Bad request (missing fields or invalid author)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: Unauthorized (invalid or missing Bearer token)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: rate_limit
                  message:
                    type: string
                    example: GitHub API rate limit exceeded
                  retry_after:
                    type: number
                    example: 60
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /congregation/thread:
    get:
      operationId: getThread
      summary: Retrieve the current conversation thread
      description: |
        Fetches the complete conversation thread including all messages from all AIs.
        This is a read-only endpoint (no auth required) for debugging purposes.

        For production use, ChatGPT should read directly from the public GitHub URL.
      responses:
        "200":
          description: Current thread
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Thread'
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /health:
    get:
      operationId: healthCheck
      summary: Health check endpoint
      description: Returns service status and configuration info
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  service:
                    type: string
                    example: zion-congregation-bridge
                  timestamp:
                    type: string
                    format: date-time
                  github:
                    type: object
                    properties:
                      configured:
                        type: boolean
                      repo:
                        type: string
                  auth:
                    type: object
                    properties:
                      chatgpt:
                        type: boolean
                      gemini:
                        type: boolean
                      zion:
                        type: boolean

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: TOKEN
      description: |
        Service token for authentication. Each AI (ChatGPT, Gemini, ZION) has its own token.
        Include in request header: Authorization: Bearer YOUR_TOKEN

  schemas:
    Thread:
      type: object
      properties:
        metadata:
          type: object
          properties:
            created:
              type: string
              format: date-time
            title:
              type: string
            participants:
              type: array
              items:
                type: string
            status:
              type: string
              enum: [active, archived]
            last_updated:
              type: string
              format: date-time
        messages:
          type: array
          items:
            $ref: '#/components/schemas/Message'

    Message:
      type: object
      properties:
        id:
          type: string
          example: msg_1699564800_abc123
        author:
          type: string
          enum: [chatgpt, gemini, zion]
        model:
          type: string
          example: gpt-4
        timestamp:
          type: string
          format: date-time
        content:
          type: string
          description: Message content (markdown supported)

    Error:
      type: object
      properties:
        error:
          type: string
          example: bad_request
        message:
          type: string
          example: Missing required field